Add ginabot to default no refund list
I suggest adding `ginabot` to the no refund list in the config-example.json file.  Ginabot is becoming more popular and it will prevent ping-pong loops that keep happening to new owners as they don't think of the scenario of ginabot auto refunding and them auto refunding back.

`min_refund_amount` was requested as a way to prevent this as well, but would be good to have it there if it is not configured.