v2.0 - Major change how transactions are processed
This is a major change to how transactions are processed and recorded in order to fix a bug which i have been trying to track down for some time resulting in some transactions being missed completely by the software.

It is as-yet untested so please use at your own risk!